# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "HDMI Multiviewer Proxy",
    "description": "REST API for controlling UHD-401MV multiviewer via RS-232",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/audio/source": {
      "post": {
        "tags": [
          "Audio"
        ],
        "summary": "Set Audio Source",
        "description": "Set audio source.\n\n- 0: Follow window 1 selected source\n- 1: HDMI 1 input audio\n- 2: HDMI 2 input audio\n- 3: HDMI 3 input audio\n- 4: HDMI 4 input audio",
        "operationId": "set_audio_source_api_audio_source_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAudioSourceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/audio/volume": {
      "post": {
        "tags": [
          "Audio"
        ],
        "summary": "Set Audio Volume",
        "description": "Set audio volume (0-100).",
        "operationId": "set_audio_volume_api_audio_volume_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetVolumeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/audio/mute": {
      "post": {
        "tags": [
          "Audio"
        ],
        "summary": "Set Audio Mute",
        "description": "Set audio mute state.",
        "operationId": "set_audio_mute_api_audio_mute_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetMuteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AudioResponse": {
        "properties": {
          "source": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source"
          },
          "source_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Name"
          },
          "volume": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Volume"
          },
          "muted": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Muted"
          }
        },
        "type": "object",
        "title": "AudioResponse",
        "description": "Audio state response."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "SetAudioSourceRequest": {
        "properties": {
          "source": {
            "type": "integer",
            "maximum": 4.0,
            "minimum": 0.0,
            "title": "Source",
            "description": "0=follow window 1, 1-4=HDMI inputs"
          }
        },
        "type": "object",
        "required": [
          "source"
        ],
        "title": "SetAudioSourceRequest",
        "description": "Set audio source request."
      },
...
      "SetMuteRequest": {
        "properties": {
          "muted": {
            "type": "boolean",
            "title": "Muted"
          }
        },
        "type": "object",
        "required": [
          "muted"
        ],
        "title": "SetMuteRequest",
        "description": "Set mute request."
      },
...
      "SetVolumeRequest": {
        "properties": {
          "volume": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Volume"
          }
        },
        "type": "object",
        "required": [
          "volume"
        ],
        "title": "SetVolumeRequest",
        "description": "Set volume request."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}